feat(DTFS2-6794): partially refactor portal users
